Beginnings on a variation of the Factory pattern
Initial implementation of a variation of the Factory pattern. There is
no Factory class and instead the loading of the required classes and
their instantiations are done dynamically. The classes that need to be
instantiated are specified in a configuration file. This way, the main
loop does not have to be altered when additional functionality is
needed. A new class can be created that implements the required
functionality and then its usages is regulated through a configuration
file, including the values that the class needs at instantiation.