Simplify the EventHandler API.
Instead of adding/deletling users to/from a group with the cua group
type in mind, simply add a user to a group and let a concrete class
worry about about the group type if there is any. This way, the main
loop is freed of this clutter (group type is a cua feature) and is put
in the right place, i.e. the concrete class.