Be more robust by writing provisional status file
The sync script assumes that a generated script has been run and without
errors. Thus that the destination LDAP has been synced properly.
However, this is not necessarily the case. In stead of just dumping the
status file and hope for the best, a provisional status file can be used
instead. When the generated scripts fails for any reason, the original
status file is still there. Thus all the previous changes will be
replayed and saved from being lost. When the generated script finished
correctly it should replace the status file with the provisional one.